% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CCPredict.R
\name{generate.test.inxs}
\alias{generate.test.inxs}
\title{generate.test.inxs}
\usage{
generate.test.inxs(n, kfold)
}
\arguments{
\item{n}{blah blah}

\item{kfold}{blah blah}
}
\value{
blah blah
}
\description{
blah blah blah
}
\examples{
X <- read.csv(system.file("extdata", "X.csv", package="CCPredict"),header=FALSE)
}

