% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/optimization_helpers.R
\name{optimize.cckopls}
\alias{optimize.cckopls}
\title{blah blah blah}
\usage{
optimize.cckopls(X, ytr, L, noxRange, LambdaRange, kfold)
}
\arguments{
\item{X}{blah blah}

\item{ytr}{blah blah}

\item{L}{blah blah}

\item{noxRange}{blah blah}

\item{LambdaRange}{blah blah}

\item{kfold}{blah blah}
}
\value{
blah blah
}
\description{
blah blah blah
}
\examples{
a=1
}

