% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CCPredict.R
\name{optimize.ccSVM}
\alias{optimize.ccSVM}
\title{blah blah blah}
\usage{
optimize.ccSVM(X, ytr, L, CRange, LambdaRange, kfold = 2, cluster.size = 8)
}
\arguments{
\item{X}{blah blah}

\item{ytr}{blah blah}

\item{L}{blah blah}

\item{CRange}{blah blah}

\item{LambdaRange}{blah blah}

\item{kfold}{blah blah}
}
\value{
blah blah
}
\description{
blah blah blah
}
\examples{
X <- read.csv(system.file("extdata", "X.csv", package="CCPredict"),header=FALSE)
}

