% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CCPredict.R
\name{predict.cckopls}
\alias{predict.cckopls}
\title{blah blah blah}
\usage{
predict.cckopls(X, y, L, test.inxs, lambda, nox)
}
\arguments{
\item{X}{blah blah}

\item{y}{blah blah}

\item{L}{blah blah}

\item{test.inxs}{blah blah}

\item{lambda}{blah blah}

\item{nox}{blah blah}
}
\value{
blah blah
}
\description{
blah blah blah
}
\examples{
X <- read.csv(system.file("extdata", "X.csv", package="CCPredict"),header=FALSE)
}

