% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CCPredict.R
\name{rescaling}
\alias{rescaling}
\title{Kernel matrix rescaling}
\usage{
rescaling(X, L, lambda)
}
\arguments{
\item{X}{- input data matrix}

\item{L}{- side information matrix}

\item{lambda}{- optimized lambda value (see Li et al.)}
}
\value{
the rescaled matrices X and K
}
\description{
Uses an optimized lambda to rescale K (the kernel matrix) by way of X
(the data set)
}

