% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CCPredict.R
\name{optimize.cckopls}
\alias{optimize.cckopls}
\title{blah blah blah}
\usage{
optimize.cckopls(X, ytr, L, noxRange, LambdaRange, kfold)
}
\arguments{
\item{X}{blah blah}

\item{ytr}{blah blah}

\item{L}{blah blah}

\item{noxRange}{blah blah}

\item{LambdaRange}{blah blah}

\item{kfold}{blah blah}
}
\value{
blah blah
}
\description{
blah blah blah
}
\examples{
X <- read.csv(system.file("extdata", "X.csv", package="CCPredict"),header=FALSE)
X <- t(X)
X = scale(X,center=T,scale=T) # Scale the X data so it has a mean of 0 and a stdev of 1. Pretty standard
y <- read.csv(system.file("extdata", "y.csv", package="CCPredict"),header=FALSE)
L <- read.csv(system.file("extdata", "L.csv", package="CCPredict"),header=FALSE)
y <- as.matrix(y)
y <- factor(y[,1])
L <- as.matrix(L)
optimize.cckopls(X,y,L,c(0:3),c(1e-8,1e-4,1e-2,1,1e+2,1e+4,1e+8),5)
}

